Lambertian and Custom setup bug fixes
Bugs in the Lambertian setup and Custom profile setups are (hopefully) fixed.

Lambertian bug was that the code would overestimate how many packets should be launched at small angles due to the low angular resolution of the Lambertian_setup fitting. This has been changed to 1e-6 radians, from 1e-2 radians, or in the case of a 1st order Lambertian no fit is used at all, just an equation, sidestepping this bug.